# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Beneficiaries API",
    "description": "Backend contract for the Arabic Beneficiaries mobile application. Authenticated endpoints use an HTTP Bearer access token. Business-controlled values such as points, attendance QR values, submission status, and duplicate awards are calculated by the server.",
    "contact": { "name": "Beneficiaries application team" },
    "license": { "name": "Proprietary" },
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/v1/retreat/activities": {
      "get": {
        "tags": ["Retreat and Reflection"],
        "summary": "List active spiritual activities",
        "operationId": "list_activities_api_v1_retreat_activities_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/SpiritualActivityResponse"
                  },
                  "type": "array",
                  "title": "Response List Activities Api V1 Retreat Activities Get"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/retreat/submissions": {
      "post": {
        "tags": ["Retreat and Reflection"],
        "summary": "Submit today's retreat activities and reflection",
        "operationId": "submit_retreat_api_v1_retreat_submissions_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RetreatSubmissionRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubmissionResult" }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/retreat/reflection/latest": {
      "get": {
        "tags": ["Retreat and Reflection"],
        "summary": "Get the latest published reflection lesson",
        "operationId": "get_latest_reflection_api_v1_retreat_reflection_latest_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "anyOf": [
                    { "$ref": "#/components/schemas/ReflectionLessonResponse" },
                    { "type": "null" }
                  ],
                  "title": "Response Get Latest Reflection Api V1 Retreat Reflection Latest Get"
                }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        },
        "security": [{ "HTTPBearer": [] }]
      }
    },
    "/api/v1/retreat/reflection/{lesson_id}/complete": {
      "post": {
        "tags": ["Retreat and Reflection"],
        "summary": "Complete a reflection exercise and award points once",
        "operationId": "complete_reflection_api_v1_retreat_reflection__lesson_id__complete_post",
        "security": [{ "HTTPBearer": [] }],
        "parameters": [
          {
            "name": "lesson_id",
            "in": "path",
            "required": true,
            "schema": { "type": "string", "title": "Lesson Id" }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/SubmissionResult" }
              }
            }
          },
          "400": {
            "description": "Business rule or invalid request",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "403": {
            "description": "Authenticated but not permitted",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "404": {
            "description": "Requested resource not found",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "409": {
            "description": "Resource state conflict",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "422": {
            "description": "Field validation failed",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "429": {
            "description": "Rate or attempt limit exceeded",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          },
          "503": {
            "description": "External delivery service unavailable",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiError" }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "properties": {
          "message": { "type": "string", "title": "Message" },
          "code": { "type": "string", "title": "Code" },
          "fieldErrors": {
            "anyOf": [
              {
                "additionalProperties": { "type": "string" },
                "type": "object"
              },
              { "type": "null" }
            ],
            "title": "Fielderrors"
          }
        },
        "type": "object",
        "required": ["message", "code"],
        "title": "ApiError"
      },
...
      "ReflectionLessonResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "date": { "type": "string", "title": "Date" },
          "title": { "type": "string", "title": "Title" },
          "points": {
            "items": { "type": "string" },
            "type": "array",
            "title": "Points"
          },
          "exercise": { "type": "string", "title": "Exercise" },
          "exercisePoints": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Exercisepoints"
          },
          "completed": {
            "type": "boolean",
            "title": "Completed",
            "default": false
          }
        },
        "type": "object",
        "required": ["id", "date", "title", "points", "exercise"],
        "title": "ReflectionLessonResponse"
      },
...
      "RetreatSubmissionRequest": {
        "properties": {
          "activityIds": {
            "items": { "type": "string" },
            "type": "array",
            "maxItems": 50,
            "title": "Activityids"
          },
          "reflection": {
            "anyOf": [
              { "type": "string", "maxLength": 5000 },
              { "type": "null" }
            ],
            "title": "Reflection"
          }
        },
        "type": "object",
        "title": "RetreatSubmissionRequest"
      },
...
      "SpiritualActivityResponse": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "title": { "type": "string", "title": "Title" },
          "points": {
            "anyOf": [{ "type": "integer" }, { "type": "null" }],
            "title": "Points"
          },
          "description": {
            "anyOf": [{ "type": "string" }, { "type": "null" }],
            "title": "Description"
          }
        },
        "type": "object",
        "required": ["id", "title"],
        "title": "SpiritualActivityResponse"
      },
      "SubmissionResult": {
        "properties": {
          "id": { "type": "string", "title": "Id" },
          "pointsAwarded": { "type": "integer", "title": "Pointsawarded" },
          "totalPoints": { "type": "integer", "title": "Totalpoints" },
          "message": { "type": "string", "title": "Message" }
        },
        "type": "object",
        "required": ["id", "pointsAwarded", "totalPoints", "message"],
        "title": "SubmissionResult"
      },
...
    },
    "securitySchemes": { "HTTPBearer": { "type": "http", "scheme": "bearer" } }
  }
}